requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.4.0
djangorestframework==3.14.0
pillow==10.4.0
pytz==2024.1
sqlparse==0.5.1
tzdata==2024.1
